Add database to cache animals for faster retrieval inside details screen

+++ app/src/main/java/com/stylianosgakis/pawfinder/ui/NavigationActions.kt
@@ -18,7 +18,7 @@ package com.stylianosgakis.pawfinder.ui
 import androidx.navigation.NavHostController
 
 class NavigationActions(navController: NavHostController) {
-    val goToDetailsScreen: (Int) -> Unit = { id ->
+    val goToDetailsScreen: (Long) -> Unit = { id ->
         if (navController.currentDestination?.route != Screen.DetailsScreen.route) {
             navController.navigate(Screen.DetailsScreen.createRoute(id))
         }
